Item functionality in battle and Item shop
Added the functionality to modify player status in battle based on equipped items.
Also added a fully functional item shop (the price will need to be edited, right now it is set to 0 in order to test, and there are no sprites for the items)

The only missing piece is saving the bought and equiped equipment to the disk and changing shop so that you can't buy them again if you already bought them.

+++ ClickUpAdventurers/Assets/_ClickUpAdventurers/Scripts/Characters/Players/Looter.cs
@@ -10,7 +10,7 @@ namespace ClickUpAdventurers
     {
         public float gatherTime = 1.0f; //The time it takes to gather 1 loot
         public int capacity = 5;        //The maximum capacity, once reached he will return to the party
-        public int movementSpeed = 5;
+        public float movementSpeed = 5;
         public int health = 3;
 
         public Image healthbar;
